[Day 3] Added solution also for part 2 of day 3
Note that the second part solution has been done in a different way compared to the first part (where a direct char-by-char comparison is made). The part 1 can be simplified by parsing also other characters in addition to "gears", I leave it in the old way to compare the two different approaches.